cas: dir: implement v1.0.0-rc5 on-disk format
The main changes in v1.0.0-rc5 was the introduction of index.json for
storing references using annotations. While there is still some upstream
discussion about certain edge-cases and how ImageIndex should be handled
in general, the general idea has already been enshrined in the spec.

There are also some changes to media-types, but those are not really
critical changes.

Signed-off-by: Aleksa Sarai <asarai@suse.de>